Work around for repeated viewer copyright.
This happens when autobuild outputs everything twice.
I can't figure out why it does that, but it does.

The problem of that is that the second 'viewer copyright'
line is then appended to the copyright line(s) of the last
packages, causing the error "Duplicated copyright" (thinking
we have two packages with the same name but a different
copyright).