Use PROPERTY POSITION_INDEPENDENT_CODE where needed (and not add a -fPIC).
A *lot* of static archives were compiled without -fPIC on linux
while they are linked against plugins, which are shared libraries.

It is possible that the GNU linker doesn't complain about this because
all x86_64 assembly is "PIC" by default (I think), but using a faster
linker like mold reveals this inconsistency.